+++ src/components/user-navigation/user-navigation-list.tsx
@@ -1,19 +1,17 @@
 import { AuthorizationStatus } from '../../const';
 import { useAppSelector } from '../../hooks';
-import UserLogout from './user-login';
 import UserProfileAuth from './user-profile-auth';
 import UserProfileNoAuth from './user-profile-no-auth';
 
-
 function HeaderNavList(): JSX.Element {
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
+  const isAuth = authorizationStatus === AuthorizationStatus.Auth;
 
-  return (
-    <ul className="header__nav-list">
-      {authorizationStatus === AuthorizationStatus.Auth ? <UserProfileAuth/> : <UserProfileNoAuth/>}
-      {authorizationStatus === AuthorizationStatus.Auth ? <UserLogout/> : ''}
-    </ul>
-  );
+  if (isAuth) {
+    return (<UserProfileAuth/>);
+  } else {
+    return (<UserProfileNoAuth/>);
+  }
 }
 
 export default HeaderNavList;
