+++ src/components/user-navigation/user-profile-auth.tsx
@@ -1,25 +1,43 @@
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const';
-import { useAppSelector } from '../../hooks';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { logoutAction } from '../../store/api-actions';
 
 function UserProfileAuth(): JSX.Element {
   const userEmail = useAppSelector((state) => state.email);
+  const dispatch = useAppDispatch();
+
+  const handleLogoutClick = (evt: React.MouseEvent<HTMLAnchorElement>) => {
+    evt.preventDefault();
+    dispatch(logoutAction());
+  };
 
   return (
-    <li className="header__nav-item user">
-      <Link
-        className="header__nav-link header__nav-link--profile"
-        to={AppRoute.Favorites}
-      >
-        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-        <span className="header__user-name user__name">
-          {userEmail}
-        </span>
-        <span className="header__favorite-count">
+    <ul className="header__nav-list">
+      <li className="header__nav-item user">
+        <Link
+          className="header__nav-link header__nav-link--profile"
+          to={AppRoute.Favorites}
+        >
+          <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+          <span className="header__user-name user__name">
+            {userEmail}
+          </span>
+          <span className="header__favorite-count">
           3
-        </span>
-      </Link>
-    </li>
+          </span>
+        </Link>
+      </li>
+      <li className="header__nav-item">
+        <Link
+          className="header__nav-link"
+          to={AppRoute.Main}
+          onClick={handleLogoutClick}
+        >
+          <span className="header__signout">Sign out</span>
+        </Link>
+      </li>
+    </ul>
   );
 }
