+++ src/components/user-navigation/user-profile-no-auth.tsx
@@ -4,15 +4,17 @@ import { AppRoute } from '../../const';
 function UserProfileNoAuth(): JSX.Element {
 
   return (
-    <li className="header__nav-item user">
-      <Link
-        className="header__nav-link header__nav-link--profile"
-        to={AppRoute.Login}
-      >
-        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-        <span className="header__login">Sign in</span>
-      </Link>
-    </li>
+    <ul className="header__nav-list">
+      <li className="header__nav-item user">
+        <Link
+          className="header__nav-link header__nav-link--profile"
+          to={AppRoute.Login}
+        >
+          <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+          <span className="header__login">Sign in</span>
+        </Link>
+      </li>
+    </ul>
   );
 }
